# Speedup compilation of substrate-network-libp2p