Before:
```
$ touch 
$ cargo check -p substrate-network-libp2p
   Finished dev [unoptimized + debuginfo] target(s) in 4m 38s
$ cargo build -p substrate-network-libp2p
   Finished dev [unoptimized + debuginfo] target(s) in 8m 42s
```

After:
```
$ cargo check -p substrate-network-libp2p
   Finished dev [unoptimized + debuginfo] target(s) in 1.85s
$ cargo build -p substrate-network-libp2p
   Finished dev [unoptimized + debuginfo] target(s) in 36.97s
```

CI runs in 12 minutes instead of 27